Libarify `identity` service
* `identity`: Libarify
    * `main.rs`:
    	* extract libary functionality into `lib.rs`
	* extract server startup into `rpc::mod.rs`
    * `Cargo.toml`: adapt libary configuration
    * `rpc::service`: Extract models into `rpc::models`
* `api::{rpc, endpoints::identity}`: Adapt to identity lib restructuring
