+++ identity/src/rpc/service.rs
@@ -1,6 +1,4 @@
-use serde::{Deserialize, Serialize};
-
-pub use helpers::rpc::{Error, RpcResult};
+use super::models::*;
 
 #[tarpc::service]
 pub trait IdentityService {
@@ -20,52 +18,4 @@ pub trait IdentityService {
     async fn oauth_client_identifier() -> RpcResult<OauthClientIdentifier>;
     async fn oauth_authentication(code: AuthorizationCode) -> RpcResult<SessionToken>;
     async fn session_info(token: String) -> RpcResult<SessionInfo>;
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct User {
-    pub id: i32,
-    pub sub: String,
-    pub email: String,
-    pub given_name: String,
-    pub family_name: String,
-    pub picture: String,
-    pub active: bool,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct Role {
-    pub id: i32,
-    pub name: String,
-    pub access_manage_books: bool,
-    pub access_manage_roles: bool,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct UserRole {
-    pub id: i32,
-    pub user_id: i32,
-    pub role_id: i32,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct OauthClientIdentifier {
-    pub identifier: String,
-}
-
-pub type AuthorizationCode = String;
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct SessionToken {
-    pub token: String,
-}
-
-#[derive(Debug, Deserialize, Serialize)]
-pub struct SessionInfo {
-    pub sub: u32,
-    pub given_name: String,
-    pub family_name: String,
-    pub picture: String,
-    pub iat: u64,
-    pub exp: u64,
 }
